moving jest config files
