+++ test/jest/setup.global.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const removeTmpPath = require('./test/helpers/remove-temp-path.js');
+const removeTmpPath = require('../helpers/remove-temp-path.js');
 
 module.exports = async() => {
     await removeTmpPath();
